// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A date-time string at UTC, such as 2007-12-03T10:15:30Z,
   *     compliant with the 'date-time' format outlined in section 5.6 of
   *     the RFC 3339 profile of the ISO 8601 standard for representation
   *     of dates and times using the Gregorian calendar.
   */
  DateTime: any;
  /** The 'Dimension' type represents dimensions as whole numeric values between `1` and `4000`. */
  Dimension: any;
  /** The 'HexColor' type represents color in `rgb:ffffff` string format. */
  HexColor: any;
  /** The 'Quality' type represents quality as whole numeric values between `1` and `100`. */
  Quality: any;
};

/** Represents a binary file in a space. An asset can be any file type. */
export type Asset = {
  __typename?: 'Asset';
  contentType?: Maybe<Scalars['String']>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<Scalars['String']>;
  fileName?: Maybe<Scalars['String']>;
  height?: Maybe<Scalars['Int']>;
  linkedFrom?: Maybe<AssetLinkingCollections>;
  size?: Maybe<Scalars['Int']>;
  sys: Sys;
  title?: Maybe<Scalars['String']>;
  url?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
};
// ...
export type AssetCollection = {
  __typename?: 'AssetCollection';
  items: Array<Maybe<Asset>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Card = Entry & {
  __typename?: 'Card';
  avatar?: Maybe<Asset>;
  contentfulMetadata: ContentfulMetadata;
  cv?: Maybe<Entry>;
  linkedFrom?: Maybe<CardLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  role?: Maybe<Scalars['String']>;
  sys: Sys;
};
// ...
export type CardCollection = {
  __typename?: 'CardCollection';
  items: Array<Maybe<Card>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type CardLinkingCollections = {
  __typename?: 'CardLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type ContentfulMetadata = {
  __typename?: 'ContentfulMetadata';
  tags: Array<Maybe<ContentfulTag>>;
};
// ...
export type Cv = Entry & {
  __typename?: 'Cv';
  age?: Maybe<Scalars['Int']>;
  basic?: Maybe<Array<Maybe<Scalars['String']>>>;
  contentfulMetadata: ContentfulMetadata;
  dateOfBirth?: Maybe<Scalars['DateTime']>;
  educationCollection?: Maybe<CvEducationCollection>;
  expertise?: Maybe<Array<Maybe<Scalars['String']>>>;
  gender?: Maybe<Scalars['String']>;
  interestsCollection?: Maybe<CvInterestsCollection>;
  knowledge?: Maybe<Array<Maybe<Scalars['String']>>>;
  languages?: Maybe<Array<Maybe<Scalars['String']>>>;
  linkedFrom?: Maybe<CvLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  photo?: Maybe<Asset>;
  professionalExperienceCollection?: Maybe<CvProfessionalExperienceCollection>;
  role?: Maybe<Scalars['String']>;
  skills?: Maybe<Array<Maybe<Scalars['String']>>>;
  sys: Sys;
};
// ...
export type CvCollection = {
  __typename?: 'CvCollection';
  items: Array<Maybe<Cv>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Education = Entry & {
  __typename?: 'Education';
  contentfulMetadata: ContentfulMetadata;
  endDate?: Maybe<Scalars['DateTime']>;
  linkedFrom?: Maybe<EducationLinkingCollections>;
  school?: Maybe<Scalars['String']>;
  startDate?: Maybe<Scalars['DateTime']>;
  studyField?: Maybe<Scalars['String']>;
  sys: Sys;
  title?: Maybe<Scalars['String']>;
};
// ...
export type EducationCollection = {
  __typename?: 'EducationCollection';
  items: Array<Maybe<Education>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Entry = {
  contentfulMetadata: ContentfulMetadata;
  sys: Sys;
};

export type EntryCollection = {
  __typename?: 'EntryCollection';
  items: Array<Maybe<Entry>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Experience = Entry & {
  __typename?: 'Experience';
  company?: Maybe<Scalars['String']>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<Scalars['String']>;
  endDate?: Maybe<Scalars['DateTime']>;
  linkedFrom?: Maybe<ExperienceLinkingCollections>;
  startDate?: Maybe<Scalars['DateTime']>;
  sys: Sys;
  title?: Maybe<Scalars['String']>;
};
// ...
export type ExperienceCollection = {
  __typename?: 'ExperienceCollection';
  items: Array<Maybe<Experience>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Interest = Entry & {
  __typename?: 'Interest';
  contentfulMetadata: ContentfulMetadata;
  icon?: Maybe<Asset>;
  linkedFrom?: Maybe<InterestLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  sys: Sys;
};
// ...
export type InterestCollection = {
  __typename?: 'InterestCollection';
  items: Array<Maybe<Interest>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type PharmacyNetwork = Entry & {
  __typename?: 'PharmacyNetwork';
  contentfulMetadata: ContentfulMetadata;
  linkedFrom?: Maybe<PharmacyNetworkLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  sys: Sys;
};
// ...
export type PharmacyNetworkCollection = {
  __typename?: 'PharmacyNetworkCollection';
  items: Array<Maybe<PharmacyNetwork>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Query = {
  __typename?: 'Query';
  asset?: Maybe<Asset>;
  assetCollection?: Maybe<AssetCollection>;
  card?: Maybe<Card>;
  cardCollection?: Maybe<CardCollection>;
  cv?: Maybe<Cv>;
  cvCollection?: Maybe<CvCollection>;
  education?: Maybe<Education>;
  educationCollection?: Maybe<EducationCollection>;
  entryCollection?: Maybe<EntryCollection>;
  experience?: Maybe<Experience>;
  experienceCollection?: Maybe<ExperienceCollection>;
  interest?: Maybe<Interest>;
  interestCollection?: Maybe<InterestCollection>;
  pharmacyNetwork?: Maybe<PharmacyNetwork>;
  pharmacyNetworkCollection?: Maybe<PharmacyNetworkCollection>;
};
// ...
export type Sys = {
  __typename?: 'Sys';
  environmentId: Scalars['String'];
  firstPublishedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['String'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  publishedVersion?: Maybe<Scalars['Int']>;
  spaceId: Scalars['String'];
};
// ...
export type AssetFragment = { __typename?: 'Asset', title?: string | null, description?: string | null, contentType?: string | null, fileName?: string | null, size?: number | null, url?: string | null, width?: number | null, height?: number | null } & { ' $fragmentName'?: 'AssetFragment' };

export type CardFragment = { __typename?: 'Card', name?: string | null, role?: string | null, avatar?: (
    { __typename?: 'Asset' }
    & { ' $fragmentRefs'?: { 'AssetFragment': AssetFragment } }
  ) | null, cv?: { __typename?: 'Card', sys: { __typename?: 'Sys', id: string } } | { __typename?: 'Cv', sys: { __typename?: 'Sys', id: string } } | { __typename?: 'Education', sys: { __typename?: 'Sys', id: string } } | { __typename?: 'Experience', sys: { __typename?: 'Sys', id: string } } | { __typename?: 'Interest', sys: { __typename?: 'Sys', id: string } } | { __typename?: 'PharmacyNetwork', sys: { __typename?: 'Sys', id: string } } | null } & { ' $fragmentName'?: 'CardFragment' };

export type GetCardsQueryVariables = Exact<{
  limit?: InputMaybe<Scalars['Int']>;
  preview?: InputMaybe<Scalars['Boolean']>;
}>;


export type GetCardsQuery = { __typename?: 'Query', cardCollection?: { __typename?: 'CardCollection', items: Array<(
      { __typename?: 'Card' }
      & { ' $fragmentRefs'?: { 'CardFragment': CardFragment } }
    ) | null> } | null };
// ...
export const AssetFragmentDoc = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"Asset"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Asset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"contentType"}},{"kind":"Field","name":{"kind":"Name","value":"fileName"}},{"kind":"Field","name":{"kind":"Name","value":"size"}},{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}}]}}]} as unknown as DocumentNode<AssetFragment, unknown>;
export const CardFragmentDoc = {"kind":"Document", "definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"Card"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Card"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"avatar"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"Asset"}}]}},{"kind":"Field","name":{"kind":"Name","value":"role"}},{"kind":"Field","name":{"kind":"Name","value":"cv"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"sys"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}}]}}]}}]}},...AssetFragmentDoc.definitions]} as unknown as DocumentNode<CardFragment, unknown>;
// ...
export const GetCardsDocument = {"kind":"Document", "definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetCards"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"preview"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"cardCollection"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}},{"kind":"Argument","name":{"kind":"Name","value":"preview"},"value":{"kind":"Variable","name":{"kind":"Name","value":"preview"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"Card"}}]}}]}}]}},...CardFragmentDoc.definitions]} as unknown as DocumentNode<GetCardsQuery, GetCardsQueryVariables>;